+++ src/main/java/DAO/DAO.java
@@ -4,15 +4,15 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface DAO<T> {
-    T get(int id) throws SQLException;
+    T get(Long id) throws SQLException;
 
     List<T> getAll() throws SQLException;
 
-    int save(T t) throws SQLException;
+    Long save(T t) throws SQLException;
 
     Long insert(T t) throws SQLException;
 
-    int update(T t) throws SQLException;
+    Long update(T t) throws SQLException;
 
-    int delete(T t);
+    Long delete(T t);
 }
